Rename Puna files
