+++ TombEngine/TombEngine.vcxproj
@@ -167,7 +167,7 @@ CALL gen.bat</Command>
     <ClInclude Include="Game\GuiObjects.h" />
     <ClInclude Include="Objects\TR1\Trap\DamoclesSword.h" />
     <ClInclude Include="Objects\TR3\Entity\Lizard.h" />
-    <ClInclude Include="Objects\TR3\Entity\tr3_punaboss.h" />
+    <ClInclude Include="Objects\TR3\Entity\PunaBoss.h" />
     <ClInclude Include="Objects\TR3\Entity\tr3_willard.h" />
     <ClInclude Include="Objects\TR3\Object\tr3_boss_object.h" />
     <ClInclude Include="Scripting\Internal\TEN\Effects\EffectIDs.h" />
@@ -640,7 +640,7 @@ CALL gen.bat</Command>
     <ClCompile Include="Game\GuiObjects.cpp" />
     <ClCompile Include="Objects\TR1\Trap\DamoclesSword.cpp" />
     <ClCompile Include="Objects\TR3\Entity\Lizard.cpp" />
-    <ClCompile Include="Objects\TR3\Entity\tr3_punaboss.cpp" />
+    <ClCompile Include="Objects\TR3\Entity\PunaBoss.cpp" />
     <ClCompile Include="Objects\TR3\Object\tr3_boss_object.cpp" />
     <ClCompile Include="Scripting\Internal\TEN\Objects\Lara\LaraObject.cpp" />
     <ClCompile Include="Scripting\Internal\TEN\Objects\Room\RoomObject.cpp" />
